feat: added spring security

+++ src/main/java/me/akshawop/journalApp/controller/AdminUserController.java
@@ -1,12 +1,13 @@
 package me.akshawop.journalApp.controller;
 
+import java.util.List;
+
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
-import org.springframework.security.core.Authentication;
-import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PutMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
@@ -16,28 +17,40 @@ import me.akshawop.journalApp.entity.User;
 import me.akshawop.journalApp.service.UserService;
 
 @RestController
-@RequestMapping("/user")
-public class UserController {
+@RequestMapping("/admin/user")
+public class AdminUserController {
     @Autowired
     private UserService service;
 
     @GetMapping
-    public ResponseEntity<User> getUserDetails() {
+    public ResponseEntity<List<User>> getAllUsers() {
         try {
-            Authentication auth = SecurityContextHolder.getContext().getAuthentication();
-            String username = auth.getName();
+            List<User> users = service.getAllUsers();
+            if (users.isEmpty())
+                return new ResponseEntity<>(HttpStatus.NO_CONTENT);
+            return new ResponseEntity<>(users, HttpStatus.OK);
+        } catch (Exception e) {
+            return new ResponseEntity<>(HttpStatus.INTERNAL_SERVER_ERROR);
+        }
+    }
+
+    @GetMapping("/{username}")
+    public ResponseEntity<User> getUserByUsername(@PathVariable String username) {
+        try {
             User user = service.getUserByUsername(username);
+            if (user == null)
+                return new ResponseEntity<>(HttpStatus.NOT_FOUND);
             return new ResponseEntity<>(user, HttpStatus.OK);
         } catch (Exception e) {
             return new ResponseEntity<>(HttpStatus.INTERNAL_SERVER_ERROR);
         }
     }
 
-    @DeleteMapping
-    public ResponseEntity<?> deleteUser() {
+    @DeleteMapping("/{username}")
+    public ResponseEntity<?> deleteUser(@PathVariable String username) {
         try {
-            Authentication auth = SecurityContextHolder.getContext().getAuthentication();
-            String username = auth.getName();
+            if (service.getUserByUsername(username) == null)
+                return new ResponseEntity<>(HttpStatus.NOT_FOUND);
 
             service.deleteUserByUsername(username);
             return new ResponseEntity<>(HttpStatus.OK);
@@ -46,14 +59,16 @@ public class UserController {
         }
     }
 
-    @PutMapping
-    public ResponseEntity<?> updateUser(@RequestBody User user) {
+    @PutMapping("/{username}")
+    public ResponseEntity<?> updateUser(@PathVariable String username, @RequestBody User user) {
         try {
-            Authentication auth = SecurityContextHolder.getContext().getAuthentication();
-            String username = auth.getName();
             User userInDb = service.getUserByUsername(username);
 
-            service.updateUser(userInDb, user);
+            // checks if the said entry exists
+            if (userInDb == null)
+                return new ResponseEntity<>(HttpStatus.NOT_FOUND);
+
+            service.updateUserAdmin(userInDb, user);
             return new ResponseEntity<>(HttpStatus.OK);
         } catch (NullPointerException e) {
             return new ResponseEntity<>(HttpStatus.BAD_REQUEST);
